+++ app/srcs/Grid.cpp
@@ -6,7 +6,7 @@
 /*   By: craimond <bomboclat@bidol.juis>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/08 15:00:10 by craimond          #+#    #+#             */
-/*   Updated: 2024/06/13 20:14:13 by craimond         ###   ########.fr       */
+/*   Updated: 2024/06/14 13:38:02 by craimond         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,13 +18,13 @@
 
 using std::array;
 
-Grid::Grid(const int32_t n_cols, const int32_t n_rows, const enum e_cell_type status)
+Grid::Grid(const uint16_t n_cols, const uint16_t n_rows, const enum e_cell_type status)
 {
 	_grid.resize(n_cols);
-	for (int32_t x = 0; x < n_cols; x++)
+	for (uint16_t x = 0; x < n_cols; x++)
 	{
 		_grid[x].resize(n_rows);
-		for (int32_t y = 0; y < n_rows; y++)
+		for (uint16_t y = 0; y < n_rows; y++)
 			_grid[x][y] = new Node(status, x, y);
 	}
 }
@@ -36,10 +36,10 @@ Grid::~Grid(void)
 			delete cell;
 }
 
-Cell		&Grid::operator()(const int32_t x, const int32_t y) { return *_grid[x][y]; }
-const Cell	&Grid::operator()(const int32_t x, const int32_t y) const { return *_grid[x][y]; }
-int32_t		Grid::getCols(void) const { return _grid.size(); }
-int32_t		Grid::getRows(void) const { return _grid.empty() ? 0 : _grid[0].size(); }
+Cell		&Grid::operator()(const uint16_t x, const uint16_t y) { return *_grid[x][y]; }
+const Cell	&Grid::operator()(const uint16_t x, const uint16_t y) const { return *_grid[x][y]; }
+uint16_t	Grid::getCols(void) const { return _grid.size(); }
+uint16_t	Grid::getRows(void) const { return _grid.empty() ? 0 : _grid[0].size(); }
 
 Cell	&Grid::getStart(void) const
 {
@@ -61,14 +61,13 @@ Cell	&Grid::getEnd(void) const
 
 float	Grid::computeDistance(const Cell &a, const Cell &b)
 {
-	const Vector2D<int32_t>	&pos_a = a.getPos();
-	const Vector2D<int32_t>	&pos_b = b.getPos();
-	const int32_t			dx = ::abs(pos_a.x - pos_b.x);
-	const int32_t			dy = ::abs(pos_a.y - pos_b.y);
+	const Vector2D<uint16_t>	&pos_a = a.getPos();
+	const Vector2D<uint16_t>	&pos_b = b.getPos();
+	const uint16_t				dx = ::abs(pos_a.x - pos_b.x);
+	const uint16_t				dy = ::abs(pos_a.y - pos_b.y);
 
 	return sqrt(dx * dx + dy * dy);
 }
-//TODO fare heuristic
 
 void	Grid::reset(void)
 {
@@ -79,7 +78,7 @@ void	Grid::reset(void)
 
 void	Grid::insertCell(const Cell &cell)
 {
-	const Vector2D<int32_t>	&pos = cell.getPos();
+	const Vector2D<uint16_t>	&pos = cell.getPos();
 
 	delete _grid[pos.x][pos.y];
 	_grid[pos.x][pos.y] = new Node(cell);
