+++ app/srcs/headers/constants.hpp
@@ -6,7 +6,7 @@
 /*   By: craimond <bomboclat@bidol.juis>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/07 18:49:01 by craimond          #+#    #+#             */
-/*   Updated: 2024/06/13 20:44:57 by craimond         ###   ########.fr       */
+/*   Updated: 2024/06/14 12:45:03 by craimond         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -24,8 +24,8 @@ constexpr uint8_t						FPS = 60;
 constexpr uint8_t						TILE_SIZE = 20;
 constexpr uint8_t						N_COLS = 40;
 constexpr uint8_t						N_ROWS = 25;
-constexpr int32_t						WIN_WIDTH = N_COLS * TILE_SIZE;
-constexpr int32_t						WIN_HEIGHT = N_ROWS * TILE_SIZE;
+constexpr uint16_t						WIN_WIDTH = N_COLS * TILE_SIZE;
+constexpr uint16_t						WIN_HEIGHT = N_ROWS * TILE_SIZE;
 extern const std::array<sf::Color, 255>	GRADIENTS;
 
 const std::array<sf::Color, 255>		precompute_gradients(void);
